fix: 完成多批次bug修复与功能优化
涵盖硬件互斥、按键修饰符计数、ASR线程安全、配置持久化、频谱归一化、OTA超时、配置原子写入等多项修复

+++ src/protocol/commands/cmd_pc_status.cpp
@@ -48,8 +48,9 @@ namespace ekeys::protocol::commands
                     Configuration::instance().mutateSettings(
                         [&mask](DeviceSettings &d)
                         { d.pc_status_mask = mask; });
+                    /* F6 修复：用 uint32 重载，高位不再被截断 */
                     Configuration::instance().saveSetting(
-                        "pc_status_mask", static_cast<int>(mask));
+                        "pc_status_mask", mask);
                 }
                 SerialProtocol::instance().sendSuccessResponse(cmd, seq,
                                                                JsonObject());
